Move stat datum formatting out of component and into stat util to allow re-use for text stats
